requirements.txt
addict==2.4.0
appnope==0.1.3
apptools==5.2.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
bounded-pool-executor==0.0.3
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
ConfigArgParse==1.5.3
configobj==5.0.8
contourpy==1.0.7
cycler==0.11.0
dash==2.10.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
envisage==7.0.3
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.0
Flask==2.2.3
fonttools==4.39.4
fsspec==2023.5.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipdb==0.13.13
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_core==5.3.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mayavi==4.8.1
mpmath==1.3.0
nbformat==5.7.0
networkx==3.1
numpy==1.24.3
open3d==0.17.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.14.1
pqdm==0.2.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pyface==8.0.0
Pygments==2.15.1
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
six==1.16.0
stack-data==0.6.2
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
traits==6.4.1
traitsui==8.0.0
transformers==4.29.2
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
vtk==9.2.6
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.15.0
